requirements.txt
aiogram>=3.0.0
supabase>=2.2.0
python-dotenv>=1.0.0
pydantic>=2.0.0
